refactor(runtime): migrate js helpers to typescript extension modules
- Move api_handler.js to api_handler.ts in runtime/ext/rari
- Move component_loader_helper.js to component_loader.ts in runtime/ext/rari
- Move cookies_helper.js to cookies.ts in runtime/ext/rari
- Move metadata_collector_helper.js to metadata_collector.ts in runtime/ext/rari
- Add TypeScript type definitions (types.d.ts) for runtime extension interfaces
- Update mod.rs to export new TypeScript modules
- Update rari.js to integrate migrated extension modules
- Update runtime_builder.rs and runtime.rs factory to reference new module locations
- Remove obsolete JavaScript helper files from factory/utils/js directory
- Consolidate runtime helpers into organized ext/rari module structure for better maintainability and type safety
